Updating price in approved barang masuk based on saldo Resolve Bug in Barang Masuk when Initialize saldo awal get from item master data Changing price to decimal with 2 digit precision Check validation on merged Barang Masuk and Moving Average Created subtotal in pengeluaran barang Found Bug in Pengeluaran Barang Cancellation (wrong updating open qty on correct permintaan) Bug in Pengeluaran Barang Cancellation Solved

+++ database/migrations/2023_08_30_034226_create_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('itemgroup_id');
             $table->string('name');
             $table->string('uom');      // 'uom' field of type string
-            $table->integer('price');   // 'price' field of type decimal with precision 
+            $table->decimal('price', 10, 2);   // 'price' field of type decimal with precision 
             $table->date('expdate')->nullable();   // 'expdate' field of type date, nullable
             $table->integer('qty')->nullable()->default(0);    // 'qty' field of type integer
             $table->integer('min_qty')->nullable()->default(0);
