+++ database/migrations/2023_09_13_064616_create_barang_masuks_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->date('docdate');
             $table->date('expdate')->nullable();   // 'expdate' field of type date, nullable
             $table->integer('qty');
-            $table->integer('price');   // 'price' field of type decimal with precision 
-            $table->integer('subtotal');
+            $table->decimal('price', 10, 2);    // 'price' field of type decimal with precision 
+            $table->decimal('subtotal', 15, 2);
             $table->string('admin');
             $table->string('po_docnum');
             $table->string('remarks');
